Improve handling of entity groups in manager
Because entities should be behaved upon differently based on their
types, there needs to be some convenient way to differentiate between
different types of entities. Even moreso because most of those had more
than one type (for example, a scientist is also a scrollable).

The way it was done so far, EntityManager had different fields for each
type of entity. Considering the nature of the type "hierarchy", this
means that managing them was manual and thus error-prone.

Now, the groups and their management was moved to a new class
EntityCollection, and each group is refered to via a corresponding tag.

A few other solutions were considered, the most notable one a custom
data strcuture of a hierarchical sets of sets of entities.
This one was chosen becauseit was deemed to have the optimal balance
between simplicity and extensibility, given the project's limited scope.